WIP(ui5-popover): Correct placement on mobile devices
Background:
The popover is positioned relative to the viewport with `position: fixed`.
However, some browsers on mobile devices have a viewport taller than the actually visible
part of the document. This problem is described here https://nicolas-hoizey.com/articles/2015/02/18/viewport-height-is-taller-than-the-visible-part-of-the-document-in-some-mobile-browsers/

Solution:
Now the popover is positioned relative to the body, and the offset is the
opener's offsetHeight.

Fixes: #3847