POC feat(ui5-list): Postpone selection of items
Background: The app developer wants to be able to put a confirmation
Dialog before the actual select of an item on user interaction.

Solution: We introduce a new property called "postponeSelection" which
when set to "true" prevents/postpones the selection of an item. The
actual selection could be made on a later state with the public method
called "resumeSelection". In case of "MultiSelect" mode, all selection
is postponed and the resumed item is the last one which was postponed.