feat(ui5-popup): Add default responsive paddings
Default responsive paddings for the content are added for all popups.
In the controls that they are not needed they are removed.
In addition for Dialog's Header and Footer are added responsive paddings.
Sample with way to remove the paddings by the user added in the test pages.

Example with removed TabContainer's content paddings added.

Fixes: #4402