feat(ui5-breadcrumbs): allow preventDefault in item-click event listener (#4589)
In a scenario where you have defined both, an item-click listener and href attributes at the breadcrumb items, one can now decide in the eventListener to not follow the link defined in href by calling event.preventDefault().
This is especially useful in SPAs with a path-based routing strategy, where you don't want to induce a page refresh when clicking on a breadcrumb item but still be able to open an items href in a new tab (e.g. by right-click on it, select open in new tab).