fix(ui5-daterange-picker): change delimiter dynamically
Problem description:
The component's input field gets cleared
when user changes the "delimiter" attribute.

Solution:
The value in the component's input field is properly
updated when the delimiter attribute is changed dynamically.

There are also two commented tests for the ui5-date-picker
component as they fail in relation to other changes.