fix(ui5-button): add new tooltip property
Problem description:
The ui5-button tooltip text was supplied trough the
private "title" property. As result the same title
attribute was placed to the ui5-button element and
to the inner native button element, which was causing
duplicate tooltip announcements.

Solution:
New public "tooltip" property is added, which
results in placing "tooltip" attribute to the ui5-button
element and a "title" attribute to the inner native button
element.

Fixes: #4689