fix(localization): fix js error if "sh" locale set (#4905)
Support for Serbian Latin has been added in UI5, now we add support for the same in UI5 Web Components.
For Serbian, there are Cyrillic and Latin scripts - "sr" maps to Cyrillic (supported prior to the change), while "sh" and "sr_latn" map to "Latin" (now supported with the current change).

FIXES: #4904