refactor(ui5-shellbar, ui5-progress-indicator, ui5-select, ui5-media-gallery): refactor RTL implementation - part 3 (#5097)
Hello @SAP/ui5-webcomponents-topic-p
this is related to the recently stopped support of IE, allowing us to implement RTL using features, not supported on IE, as the CSS logical properties.

The Browser flips the paddings, margins, etc. automatically based on the direction - there is no need to write specific RTL styles manually. At least for the styles that have CSS logical prop alternative.

In case, there is no CSS logical props alternative or the use-case is very specifi, we can extract CSS vars which will have different value in RTL and LTR and add them to global rtl-parameters.css. The same pattern (sizes-parameters.css) that we currently use for the cozy/compact styles.

Doing so, from one hand side we turn more to native implementation and standards, from the other we want to avoid effectiveDir calls as internally getComputedStyle is called, which is slow.