refactor(ui5-popups): refactor RTL implementation - part 4 (#5104)
This is related to the recently stopped support of IE, allowing us to implement RTL using features, not supported on IE, as the CSS logical properties.
The Browser flips the paddings, margins, etc. automatically based on the direction - there is no need to write specific RTL styles manually. At least for the styles that have CSS logical prop alternative.
In case, there is no CSS logical props alternative or the use-case is very specific as with the Popover's arrow margins, which are different in RTL and LTR (not even mirrored), we extract them as CSS vars and add them to global rtl-parameters.css as we do for the cozy/compact styles.
Doing so, from one hand side we turn more to native implementation and standards, from the other we want to avoid effectiveDir calls as internally getComputedStyle is called, which is slow.
Changes:
- Popover's arrow rtl styles are refactored with the rtl-parameters.css pattern
- This style does not have ltr counter part and did not apply when testing, so I removed it
```css
:host [dir="rtl"] .ui5-responsive-popover-header {
	padding: 0 1rem 0 0;
}
```
- right is replaced with inset-inline-end
```css 
.ui5-popup-resize-handle {
	/* right: var(--_ui5_dialog_resize_handle_right); */
	inset-inline-end:var(--_ui5_dialog_resize_handle_right);
}
```
- The Dialog's resize icon does not need "dir" set