fix(theming): fix fallback to default theme (#5185)
When the user sets unknown theme, the fallback does not quite work. The framework correctly displays a warning, but the theme parameters are not returned (as the theme parameters map ends up empty at that point), and the page is left in broken state. Now we use the same flow to fallback to the default theme, as if the user set it.

FIXES: #5136
Closes: #5136