fix(ui5-tabcontainer): provide unique names for each disabled slot
In Firefox, using two or more slots with the same name (i.e. "disabled-slot")
in a shadow root will prevent the other slots from being shown.
Further clicking on the Tab strip after the slots are broken will cause
the tab to crash.

This change fixes the ui5-tabcontainer to work correctly in Firefox by
ensuring each slot that should be hidden to have a unique name.

Fixes #5178