fix(ui5-dialog): reposition on screen resize
When the body element is shorter in height than the screen size,
resizing the browser window in height would not trigger a resize
event on the body, and the dialog will not be repositioned.

To solve this case, the dialog has been changed to listen to the resize
event on the window.

Fixes: #5259