fix(ui5-rating-indicator): import ui5-icon as used in template (#5320)
The Icon component is not imported, although used. Not marked as dependency.
If someone bundles just the RatingIndicator, the Icon component will be missing and actually the stars won't be displayed.
Also, when we don't mark it as dependency, we break the scoping, which adds suffixes to the dependencies of the component - the case with the issue, linked below. The users enabled the scoping feature and used ui5-rating-indicator-demo, but in the Shadow DOM, the ui5-icon did not get its suffix (ui5-icon-demo) and thus not rendered.

FIXES: #5318