fix(ui5-date-picker): adjust value help icon role
The ui5-date-picker value help icon is still decorative, but now has role "button".
On desktop or combi device the value help icon should be hidden from the screen reader.
For mobile devices or tablet the icon is visible for the screen readers, in order to enable
explore by touch scenarios.

Fixes: #5378