feat(ui5-tabcontainer): make tab-select event preventable
The `tab-select` event can now be prevented.

Added new public getter `allItems` that returns an array which includes all
slotted tabs and their subTabs. The reason for the addition is that
`tab-select` returns a `tabIndex` of the selected tab, but if that tab is
nested, then `tabIndex` will not correspond to the `items` property,
which includes only the top-level tabs.

Fixes #5116