feat: support TNT Icons v3 font (Horizon) (#6456)
Prior to this PR, we used to have a single version for SAP TNT Icons (version 2.x) and now we're adding support for new one (version 3.x), based on demand and questions from SAP Design team. Also, to keep in sync with OpenUI5, where this is also in progress. There are also changes related to the SAP Icons, explained below.

Fixes: #6497
Fixes: #6438