chore: add SAP BS icon "add-polygon" with a correct name (#6564)
Existing "add-polygone" icon has typo, reported to the SAP Design team. While this is centrally fixed, we create a copy of it, but with corrected name - "add-polygon". The old icon is still shipped, so the change is compatible. In addition, all usages of "add-polygone" are replaced with "add-polygon".
Still, in our docs we refer to the Icon Explorer where one can see only the old name, and probably won't ever see the fixed name. However, as we don't have a better alternative of the icon explorer, we will leave this as it is and wait for central fix.

Fixes: #6547