fix(ui5-select): fixed valuestate message scroll prevention (#6548)
If the valuestate message text is too long, sometimes it all the
space of the selec't popover. This way the select items are not visible
and can't be accessed by scrolling.
This is why valuestate is no longer in the header slot of the
select's popover but in the content slot. This way a scrollbar appears
when the valuestate text is too long and the select's items are not
visible/accessible.

Fixes #5970