fix(ui5-li): removed private volatile redundant event
There is a private event attached to the internal checkbox which is
being used in the "ListItem" when the its parent "List" is in
"MultiSelect" mode.
This event was previously used to tell the List that a selection is
requested (also private event), but it seems that it's no longer needed.
Even without it the selection-change in the List works properly.

The problematic event is interfering (delaying) the item-click public
event of the List. Which in the end results in different "selected"
value whether the item is clicked over the checkbox or out of it.

Fixes #6523