feat(ui5-calendar,ui5-date*picker): add static format (ISO) support for min/max dates (#6955)
We have implemented a new behavior that adds a static format for the min and max dates. This means that if a format pattern is not provided, developers can use the date part of a Date in ISO format (YYYY-MM-dd). This new behavior will help ensure that the API is more stable and easier to use.
Additionally, we have included new samples and tests to ensure compliance with this change.
Closes: #6885