feat(ui5-segmented-button): enable multiple selection
New enumeration "SegmentedButtonMode" and a new property
"mode" are introduced with allowed values:

SegmentedButtonMode.Single
SegmentedButtonMode.Multi
Multiple ui5-segmented-button-items could be selected/deselected
when the "mode" property is to "SegmentedButtonMode.Multi".
The default value for the "mode" is "SegmentedButtonMode.Single".

The "ZeroOrSingle" selection could be added in the future
if a valid use case is indicated.