feat(ui5-segmented-button): enable multiple selection
New enumeration "SegmentedButtonMode" and a new property
"mode" are introduced with allowed values:

- SegmentedButtonMode.SingleSelect
- SegmentedButtonMode.MultiSelect

Multiple ui5-segmented-button-items could be selected/deselected
when the "mode" property is to "SegmentedButtonMode.MultiSelect".
The default value for the "mode" is "SegmentedButtonMode.SingleSelect".

The "ZeroOrSingle" selection could be added in the future if a valid use case is indicated.