feat(ui5-menu): dispatch ui5-before-open for the sub-menus (#7064)
* feat(ui5-menu): dispatch ui5-before-open for the sub-menus

The ui5-before-open event is now getting fired for the
sub-menus as well. As a result the application developers
would be able to fetch data from a database before a sub-menu
gets opened. In the meanwhile the busy state of the sub-menu
will be displayed.

* feat(ui5-menu): dispatch ui5-before-open for the sub-menus

The ui5-before-open event is now getting fired for the sub-menus as well.
As a result the application developers would be able to fetch data
from a database before a sub-menu gets opened. In the meanwhile
the busy state of the sub-menu will be displayed.

* feat(ui5-menu): dispatch ui5-before-open for the sub-menus

The ui5-before-open event is now getting fired for the sub-menus as well.
As a result the application developers would be able to fetch data
from a database before a sub-menu gets opened. In the meanwhile
the busy state of the sub-menu will be displayed.

Related to: #6131

* feat(ui5-menu): dispatch ui5-before-open for the sub-menus

The ui5-before-open event is now getting fired for the sub-menus as well.
As a result the application developers would be able to fetch data from
a database before a sub-menu gets opened. In the meanwhile the busy
state of the sub-menu will be displayed.

Related to: #6131

* feat(ui5-menu): dispatch ui5-before-open for the sub-menus

The ui5-before-open event is now getting fired for the sub-menus as well.
As a result the application developers would be able to fetch data from
a database before a sub-menu gets opened. In the meanwhile the busy
state of the sub-menu will be displayed.

Related to: #6131

* feat(ui5-menu): dispatch ui5-before-open for the sub-menus

The ui5-before-open event is now getting fired for the sub-menus as well.
As a result the application developers would be able to fetch data from
a database before a sub-menu gets opened. In the meanwhile the busy
state of the sub-menu will be displayed.

Related to: #6131

* feat(ui5-menu): dispatch ui5-before-open for the sub-menus

The ui5-before-open event is now getting fired for the sub-menus as well.
As a result the application developers would be able to fetch data from
a database before a sub-menu gets opened. In the meanwhile the busy
state of the sub-menu will be displayed.

Related to: #6131

* feat(ui5-menu): dispatch ui5-before-open for the sub-menus

The ui5-before-open event is now getting fired for the sub-menus as well.
As a result the application developers would be able to fetch data from
a database before a sub-menu gets opened. In the meanwhile the busy
state of the sub-menu will be displayed.

Related to: #6131

* feat(ui5-menu): dispatch ui5-before-open for the sub-menus

The ui5-before-open event is now getting fired for the sub-menus as well.
As a result the application developers would be able to fetch data from
a database before a sub-menu gets opened. In the meanwhile the busy
state of the sub-menu will be displayed.

Related to: #6131

* feat(ui5-menu): dispatch ui5-before-open for the sub-menus

The ui5-before-open event is now getting fired for the sub-menus as well.
As a result the application developers would be able to fetch data from
a database before a sub-menu gets opened. In the meanwhile the busy
state of the sub-menu will be displayed.

Related to: #6131

* feat(ui5-menu): dispatch ui5-before-open for the sub-menus

The ui5-before-open event is now getting fired for the sub-menus as well.
As a result the application developers would be able to fetch data
from a database before a sub-menu gets opened. In the meanwhile
the busy state of the sub-menu will be displayed.

Related to: #6131