docs(ui5-file-uploader): improve accessibility showcase for samples (#7122)
As there are ongoing discussions regarding the inclusion of the new property accessibleNameRef in our ui5-file-uploader component, we have decided to enhance the accessibility aspects of the component's samples. This improvement involves providing instructions on utilizing the accessibility features by incorporating them into the examples. However, if the decision is made to include the mentioned property, the samples might need to be updated again.

Part of: #5893