fix(ui5-bar): correct padding values in dialog (#7046)
Based on the design specifications, the intended padding for both the left and right sides of the ui5-bar should be 16px. However, upon observation, it is evident that the actual padding measures only 12px ( 0.5rem padding of the endcontent container + 0.25rem margin of the slotted element ).
With this changes, the paddings are now correct.