fix(ui5-segmented-button): support custom width configuration (#6969)
With this change we enhance the UI5 SegmentedButton component to provide support for setting up custom widths. With this enhancement, you can now specify the width of the SegmentedButton using inline styles or CSS classes, ensuring that the component remains responsive in various scenarios.
Also significant reduce of JS code has been made.