style(ui5-date-picker): align value state colors with the spec
- All date and time components including the ui5-step-input
has to be aligned with the horizon theme changes.
The ui5-input component already supports those styles.
With this draft change we're trying to reuse the ui5-input
styles without copying them directly into the correspodning CSS files.
For example in the ui5-date-picker template we use ui5-input.
The styles needs to be applied on the ui5-date-picker element though.
For that reason the ui5-input styles are not imported into the date picker css.