fix(ui5-menu): fix runtime js error on `getElementById` call (#8021)
The PR adds a check to guarantee getElementById is called only if existing, e.g. the Menu is still part of the DOM tree and has not been removed in-between.
The test pages changes are not related to the change, but we have noticed that the used CSS classes "samples-margin", "header" are missing and they are not doing anything. In addition, h tags have been replaced by ui5-title, so that the texts are readable in HC themes (otherwise you get dark texts on dark background).

Fixes: #8017