feat(ui5-shellbar): introducing Joule
The coPilot is now called Joule. The coPilot keyword is kept in the code
for backwards compatibility reasons so no change of APIs and events.

The new Joule icon can be toggled and it's placed on the right of the
ShellBar (next to the search icon if present) with the rest of the
items (no longer in the center).

Additionally the search slot is now automatically placed in the center
of the shellbar on XXL breakpoint.