fix(ui5-checkbox): improved key down behavior for readonly state
Previously while holding "Space" or "Enter" key on a readonly checkbox
the "active" attribute was being applied on the checkbox main DOM node.
This was resulting in bluish tint over the readonly checkbox indicating
changing of value where this shouldn't be possible.

This wrong behavior is now fixed and "Space" or "Enter" keys down no
longer affect readonly checkbox.

Fixes: #7807