feat(ui5-shellbar): midContent private slot POC
POC of private slot named "midContent" which gives the app developer
to place an HTML element in the middle of the ShellBar.

Note that overrides the "searchField" slot and its responsive behaviors
rendering it invisible.

Minor bug needs fixing is having search icon in the OTB popover when
having both slots.