feat(ui5-menu): refactor menu items display
- This PR affects the ui5-menu and ui5-side-navigation
components rendering.
- The ui5-menu-item now extends ui5-li-custom class and it
will be represented directly as a list item in the DOM.
As a result the ui5-menu-item is no longer an abstract class,
but a physical component and it will ease the possibility
in the future to support custom content supplied by
the application developers.
- Additionally the ui5-reponsive-popover is no longer
rendered into the static area. Thus the ui5-menu structure
defined by the application developers is represented more
accurately into the DOM and the browser popover API could
be used once available for all supported browsers.

- There is also a new behavior implemented, which enables
the end users to navigate via mouse and keyboard to a
disabled ui5-menu-item element. This behavior is aimed
to improve the accessibility of the ui5-menu component
by providing screen reader announcements for disabled
menu items as well.

Fixes: #7096