refactor(framework|components): make use of `:dir(rtl)` selector in `UI5Element#getEffectiveDir` and components' styles  (#8241)
The :dir pseudo selector is now commonly supported by all browsers and the good thing about it is that ":dir() will match the value calculated by the user agent, even if inherited". To put it simpler - even if you set dir="rtl" on the body, :dir(rtl) will match in the Shadow DOM of the components.
This allows us to remove dir="{{effectiveDir}}" from the templates and replace dir[rtl] with :dir(rtl) in the component styles. Apart of making use of the standard evolution, dir="{{effectiveDir}}" is currently calling getComputedStyle and also to take effect the component needs to be re-rendered. So replacing it has performance benefits as well.
With this refactoring
- There is no usage of dir="{{effectiveDir}}" in the templates and styles relying on dir[rtl].
- The UI5Element's effectiveDir method is now based on :dir() selector and no longer calls getComputedStyle
- Remove getRTL from base configuration (it was not publicly documented)
TODO: What is left are calculations based on the effectiveDir in the component .ts files. Some of them could be potentially removed.