refactor(ui5-menu): adjust menu and sub-menu creation
- The ui5-menu elements used for sub-menus are created only once
and are being reused afterwards. They are no longer destroyed on close.
This contributes to lowering the count of the slow DOM manipulation operations.

- There is now no differentiation between mobile and desktop
device in regards to the display mechanism. In both cases
we rely on the template to do the job as the components used
for composition like ui5-list and ui5-responsive-popover do
comply with the device.

Fixes: #7767
Fixes: #7423
Fixes: #6761
Related to: #7391